♻️ MainPage의 SearchBar를 별도 컴포넌트로 분리
