+++ src/features/search/search-bar/index.ts
@@ -1,0 +1,3 @@
+import SearchBar from './SearchBar';
+
+export default SearchBar;
